Added behaviours: logging, performance and validator and modified the logic in query/commands.

+++ MyBookAPI.Application/Authors/Commands/DeleteAuthor/DeleteAuthorCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using MyBookAPI.Application.Common.Exceptions;
 using MyBookAPI.Application.Common.Interfaces;
 using System.Linq;
 using System.Threading;
@@ -18,6 +19,9 @@ namespace MyBookAPI.Application.Authors.Commands.DeleteAuthor
         {
             var author = await _context.Authors.Where(x => x.AuthorName.FirstName.Equals(request.FirstName) &&
                                                            x.AuthorName.LastName.Equals(request.LastName)).FirstOrDefaultAsync(cancellationToken);
+
+            if (author is null)
+                throw new NotFoundException($"Author {request.FirstName} {request.LastName} cannot be deleted because author does not exist.");
 
             _context.Authors.Remove(author);
             await _context.SaveChangesAsync(cancellationToken);
