+++ MyBookAPI.Application/Books/Commands/DeleteBook/DeleteBookCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using MyBookAPI.Application.Common.Exceptions;
 using MyBookAPI.Application.Common.Interfaces;
 using System.Linq;
 using System.Threading;
@@ -17,6 +18,9 @@ namespace MyBookAPI.Application.Books.Commands.DeleteBook
         public async Task<Unit> Handle(DeleteBookCommand request, CancellationToken cancellationToken)
         {
             var book = await _context.Books.Where(x => x.Name.Equals(request.BookName)).FirstOrDefaultAsync(cancellationToken);
+
+            if (book is null)
+                throw new NotFoundException($"Book {request.BookName} cannot be deleted because it does not exist.");
 
             _context.Books.Remove(book);
             await _context.SaveChangesAsync(cancellationToken);
